feat: placed test in tests folder and separated them
