+++ tests/__init__.py
@@ -1,0 +1,1 @@
+quick_check = "Some test on imports"
